Django 1.10: fix template rendering to string
Django 1.8 deprecated the `dictionary` and `context_instance` parameters
for `render_to_string()`, which have been removed in Django 1.10. This
means that instead of passing a `RequestContext` object to
`render_to_string()`, a dictionary needs to be passed in the `context`
parameter and a request in the `request` parameter.

Without this change, templates used in conjunction with
django-contact-form would have no means to access the `request` instance
in Django 1.10.